refactor: componentes structure, pages folder & app style
