+++ AppStyle.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Container = styled.view`
+    flex: 1;
+    background-color: #fff;
+    align-items: center;
+    justify-content: center;
+`;
+
+export default Container;
