+++ src/components/header/index.tsx
@@ -1,7 +1,7 @@
 import { View, Text } from 'react-native';
 import React from 'react';
 import codeMentorLogo from '../assets/code-mentor-logo.jpeg';
-import LogoImage from '../styles/header';
+import LogoImage from './style';
 
 export default function Header() {
   return (
