+++ src/pages/login-page/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './components/header';
-import LoginForm from './components/login-form';
-import LoginContainer from './styles/login-page';
+import Header from '../../components/header';
+import LoginForm from '../../components/login-form';
+import LoginContainer from './style';
 
 export default function Login() {
   return (
